somes fixes to diretorio and servidor

+++ Diretorio/src/main/java/Representation/NumberOfSick.java
@@ -5,11 +5,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class NumberOfSick {
     public final int id;
-    public final int numero;
+    public final float numero;
 
     @JsonCreator
     public NumberOfSick(@JsonProperty("Status") int id,
-                        @JsonProperty("Número médio de utilizadores que se cruzaram com utilizadores declarados doentes") int numero) {
+                        @JsonProperty("Número médio de utilizadores que se cruzaram com utilizadores declarados doentes")float numero) {
         this.id = id;
         this.numero = numero;
     }
@@ -21,7 +21,7 @@ public class NumberOfSick {
 
 
     @JsonProperty("Número médio de utilizadores que se cruzaram com utilizadores declarados doentes")
-    public int getNumero() {
+    public float getNumero() {
         return this.numero;
     }
 }
